+++ src/SYT.Fiskaly/SignDE/Exports/Dsfinvk/CashPointClosingSegment.cs
@@ -1,4 +1,0 @@
-namespace SYT.Fiskaly.SignDE.Exports.Dsfinvk;
-
-public sealed class CashPointClosingSegment(string fileName, byte[] content)
-    : DsfinvkSegment(DsfinvkSegmentType.CashPointClosing, fileName, content);