+++ src/SYT.Fiskaly/SignDE/Exports/Dsfinvk/MasterDataSegment.cs
@@ -1,4 +1,0 @@
-namespace SYT.Fiskaly.SignDE.Exports.Dsfinvk;
-
-public sealed class MasterDataSegment(string fileName, byte[] content)
-    : DsfinvkSegment(DsfinvkSegmentType.MasterData, fileName, content);