+++ src/SYT.Fiskaly/SignDE/Exports/Dsfinvk/TransactionSegment.cs
@@ -1,4 +1,0 @@
-namespace SYT.Fiskaly.SignDE.Exports.Dsfinvk;
-
-public sealed class TransactionSegment(string fileName, byte[] content)
-    : DsfinvkSegment(DsfinvkSegmentType.TransactionData, fileName, content);