+++ src/SYT.Fiskaly/SignDE/Exports/Dsfinvk/UnknownDsfinvkSegment.cs
@@ -1,4 +1,0 @@
-namespace SYT.Fiskaly.SignDE.Exports.Dsfinvk;
-
-public sealed class UnknownDsfinvkSegment(string fileName, byte[] content)
-    : DsfinvkSegment(DsfinvkSegmentType.Unknown, fileName, content);